feat: move lib folder
