+++ app/api/csrf/route.ts
@@ -1,4 +1,4 @@
-import { generateCsrfToken } from "@/app/lib/csrf"
+import { generateCsrfToken } from "@/lib/csrf"
 import { cookies } from "next/headers"
 import { NextResponse } from "next/server"
