+++ app/api/update-chat-model/route.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/app/lib/supabase/server"
+import { createClient } from "@/lib/supabase/server"
 
 export async function POST(request: Request) {
   try {
