+++ app/auth/callback/route.ts
@@ -1,6 +1,6 @@
-import { MODEL_DEFAULT } from "@/app/lib/config"
-import { createClient } from "@/app/lib/supabase/server"
-import { createGuestServerClient } from "@/app/lib/supabase/server-guest"
+import { MODEL_DEFAULT } from "@/lib/config"
+import { createClient } from "@/lib/supabase/server"
+import { createGuestServerClient } from "@/lib/supabase/server-guest"
 import { NextResponse } from "next/server"
 
 export async function GET(request: Request) {
