+++ app/auth/login/actions.ts
@@ -1,6 +1,6 @@
 "use server"
 
-import { createClient } from "@/app/lib/supabase/server"
+import { createClient } from "@/lib/supabase/server"
 import { revalidatePath } from "next/cache"
 import { redirect } from "next/navigation"
