+++ app/components/chat-input/chat-input.tsx
@@ -1,6 +1,5 @@
 "use client"
 
-import { APP_NAME } from "@/app/lib/config"
 import {
   PromptInput,
   PromptInputAction,
@@ -8,6 +7,7 @@ import {
   PromptInputTextarea,
 } from "@/components/prompt-kit/prompt-input"
 import { Button } from "@/components/ui/button"
+import { APP_NAME } from "@/lib/config"
 import { ArrowUp, Stop } from "@phosphor-icons/react"
 import React, { useCallback } from "react"
 import { ButtonFileUpload } from "./button-file-upload"
