+++ app/components/chat-input/personas.tsx
@@ -1,8 +1,8 @@
 "use client"
 
-import { PERSONAS } from "@/app/lib/config"
-import { TRANSITION_SUGGESTIONS } from "@/app/lib/motion"
 import { Button } from "@/components/ui/button"
+import { PERSONAS } from "@/lib/config"
+import { TRANSITION_SUGGESTIONS } from "@/lib/motion"
 import { cn } from "@/lib/utils"
 import { motion } from "motion/react"
 import { memo } from "react"
