+++ app/components/chat-input/suggestions.tsx
@@ -1,10 +1,10 @@
 "use client"
 
-import { TRANSITION_SUGGESTIONS } from "@/app/lib/motion"
 import { PromptSuggestion } from "@/components/prompt-kit/prompt-suggestion"
+import { TRANSITION_SUGGESTIONS } from "@/lib/motion"
 import { AnimatePresence, motion } from "motion/react"
 import React, { memo, useCallback, useEffect, useMemo, useState } from "react"
-import { SUGGESTIONS as SUGGESTIONS_CONFIG } from "../../lib/config"
+import { SUGGESTIONS as SUGGESTIONS_CONFIG } from "../../../lib/config"
 
 type SuggestionsProps = {
   onValueChange: (value: string) => void
