+++ app/components/chat/dialog-auth.tsx
@@ -1,7 +1,5 @@
 "use client"
 
-import { signInWithGoogle } from "@/app/lib/api"
-import { createClient } from "@/app/lib/supabase/client"
 import { Button } from "@/components/ui/button"
 import {
   Dialog,
@@ -11,6 +9,8 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog"
+import { signInWithGoogle } from "@/lib/api"
+import { createClient } from "@/lib/supabase/client"
 import { useState } from "react"
 
 type DialogAuthProps = {
