+++ app/layout.tsx
@@ -1,12 +1,12 @@
 import type { Metadata } from "next"
 import { Geist, Geist_Mono } from "next/font/google"
 import "./globals.css"
-import { APP_DESCRIPTION, APP_NAME } from "@/app/lib/config"
 import { Toaster } from "@/components/ui/sonner"
+import { APP_DESCRIPTION, APP_NAME } from "@/lib/config"
 import { ThemeProvider } from "next-themes"
 import Script from "next/script"
+import { createClient } from "../lib/supabase/server"
 import { LayoutClient } from "./layout-client"
-import { createClient } from "./lib/supabase/server"
 import { UserProvider } from "./providers/user-provider"
 import { UserProfile } from "./types/user"
